refactor: Use static client for blog service and move to own file
Refactors the blog data fetching functions (`getBlogPosts`, `getBlogPostBySlug`):
- Moved functions to a dedicated service file: `src/services/blog.ts`.
- Updated functions to use the server-side static PocketBase client (`createStaticClient` from `$/utils/pocketbase/static.ts`) instead of the browser client.
- Updated blog page components to import from the new service file.
- Removed old functions from `src/utils/pocketbase/client.ts`.

This change ensures that blog data fetching uses a privileged, server-side client and improves code organization.
