+++ src/utils/pocketbase/client.ts
@@ -1,6 +1,6 @@
 import { env } from '$/env';
-import PocketBase, { type ListResult } from 'pocketbase';
-import type { BlogResponse } from './pocketbase-types'; // Updated import
+import PocketBase from 'pocketbase'; // ListResult removed
+// BlogResponse import removed
 
 import { TypedPocketBase } from './pocketbase-types';
 
@@ -20,33 +20,5 @@ export function createBrowserClient() {
   return client;
 }
 
-export async function getBlogPosts(
-  page = 1,
-  limit = 10,
-): Promise<ListResult<BlogResponse>> { // Updated return type
-  const pb = createBrowserClient();
-  const result = await pb
-    .collection('blog') // Updated collection name
-    .getList<BlogResponse>(page, limit, { // Updated generic type
-      sort: '-created', // Filter removed
-    });
-  return result;
-}
-
-export async function getBlogPostBySlug(
-  slug: string,
-): Promise<BlogResponse | null> { // Updated return type
-  const pb = createBrowserClient();
-  try {
-    const post = await pb
-      .collection('blog') // Updated collection name
-      .getFirstListItem<BlogResponse>(`slug = "${slug}"`); // Updated filter and generic type
-    return post;
-  } catch (error) {
-    // PocketBase throws an error if no record is found (e.g., 404)
-    // Or if there's any other issue with the request.
-    // We'll return null to indicate not found or error.
-    console.error(`Error fetching post by slug "${slug}":`, error);
-    return null;
-  }
-}
+// getBlogPosts function removed
+// getBlogPostBySlug function removed
